MLOPS-A2/requirements.txt
dagster
dagster-webserver
dagster-mlflow
mlflow
torch
torchvision
pandas
matplotlib
Pillow
scikit-learn
numpy
seaborn
openpyxl